+++ src/modals/BatchTagModal.ts
@@ -52,7 +52,7 @@ export class BatchTagModal extends Modal {
         button
           .setButtonText("应用")
           .setCta()
-          .onClick(() => void this.submit());
+          .onClick(() => this.plugin.runAction("批量处理标签", () => this.submit()));
       })
       .addButton((button) => {
         button.setButtonText("取消").onClick(() => this.close());